Rename MyStandardScaler and fix to work with dataframes (#19)
- `MyStandardScaler` renamed to `StandardScalerWithDOF` and moved to `transformers/_base.py`
- Cast dataframe to array before setting scale_
- Expose all transformers from `transformers/__init__.py
- Add `test_estimators_support_dataframe` to ensure estimators can handle X, y dataframes
