+++ src/sknnr/_euclidean.py
@@ -1,11 +1,12 @@
 from sklearn.utils.validation import check_is_fitted
 
-from ._base import IDNeighborsRegressor, MyStandardScaler, TransformedKNeighborsMixin
+from ._base import IDNeighborsRegressor, TransformedKNeighborsMixin
+from .transformers import StandardScalerWithDOF
 
 
 class EuclideanKNNRegressor(IDNeighborsRegressor, TransformedKNeighborsMixin):
     def fit(self, X, y):
-        self.transform_ = MyStandardScaler().fit(X)
+        self.transform_ = StandardScalerWithDOF(ddof=1).fit(X)
         X = self.transform_.transform(X)
         return super().fit(X, y)
