+++ src/sknnr/_gnn.py
@@ -1,7 +1,7 @@
 from sklearn.utils.validation import check_is_fitted
 
 from ._base import IDNeighborsRegressor, TransformedKNeighborsMixin
-from .transformers._cca_transformer import CCATransformer
+from .transformers import CCATransformer
 
 
 class GNNRegressor(IDNeighborsRegressor, TransformedKNeighborsMixin):
