+++ src/sknnr/transformers/_cca_transformer.py
@@ -1,3 +1,4 @@
+import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from ._cca import CCA
@@ -6,6 +7,7 @@ from ._cca import CCA
 class CCATransformer(TransformerMixin, BaseEstimator):
     def fit(self, X, y=None, spp=None):
         y = spp if spp is not None else y
+        X, y = np.asarray(X), np.asarray(y)
         self.cca_ = CCA(X, y)
         return self
